IMPALA-9913: Use table id to match the table in drop table event
This patch modifies the matching logic of the table while
processing a DROP_TABLE event. The EventsProcessor currently relies
on createTime of the table to determine if the table in the catalog
matches with the table in the event. It is possible that if there
is a time-skew between the different HMS instances and a same table
is dropped and recreated, the creation time of the table which
is present in the catalog is less than or equals to the creationTime
of a table which has been dropped. In such case, there is a false
negative during the matching and the DROP_TABLE event is skipped.

HIVE-20556 exposes a id field of the table which can be used to
uniquely identify a table. A table which is dropped and recreated
with the same name has a different id. This patch makes use of this
id to match the table in catalog with the table from the HMS event.

Note that the DROP_DATABASE event continues to use createTime for
matching. We should use a database id for it when it is available.
I have created HIVE-23785 for tracking that on the HMS side.

Testing:
1. Ran Events processor tests.

Change-Id: Id6a80bbf5757e46318af1b57911fc127d7dd1f01
Reviewed-on: http://gerrit.cloudera.org:8080/16129
Reviewed-by: Anurag Mantripragada <anurag@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit e4a3e9a35eeb3f2791f4f80e76ca35f9a048f13b)
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/91461
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>