IMPALA-10010: Add option to configure metrics webserver
Currently, when security is turned on for the webui, eg. with
--webserver_require_ldap, authentication is applied to all webui
endpoints.

However, there are some endpoints that expose low-sensitivity info and
which are scraped by other systems that it may be difficult to get
credentials to in order to be able to authenticate.

This patch adds a flag, --metrics_webserver_port, which if specified
turns on an unsecured webserver that exposes only the /metrics,
/jsonmetrics, /metrics_prometheus for all Impala daemons. Impalads
also have the /healthz endpoint exposed.

Testing:
- Added a test that turns on the metrics server and verifies its
  reachable and unsecured.

Change-Id: Ibcf297d798a1a5c9cd59d4d82706d2d945e10d3d