IMPALA-10073: Create shaded dependency for S3A and aws-java-sdk-bundle
The aws-java-sdk-bundle is one of the largest dependencies in the Impala
Docker images and continues to grow. The jar includes SDKs for
every single AWS service.

This patch removes most of the unnecessary SDKs from the
aws-java-sdk-bundle, thus drastically decreasing the size of the
dependency. The Maven shade plugin is used to do this, and the
implementation is similar to what is currently done for the hive-exec
jar.

This patch takes a conservative approach to removing packages from the
aws-java-sdk-bundle jar, and I ensured no direct dependencies of the S3
SDK were removed. The idea is to only remove dependencies that S3A would
never conceivably need. Given the huge number of AWS services, I only
focused on removing the largest SDKs (the size of each SDK is estimated
by the number of classes in the SDK).

This decreases the size of the Docker images by about 100 MB.

Testing:
* Ran core tests against S3

Change-Id: I0939f73be986f83cc1fd07921563b4d9201780f2
Reviewed-on: http://gerrit.cloudera.org:8080/16342
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 5daff3472440dc6174f0f31a28bbdafee4f68716)
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/99096
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Abhishek Rawat <arawat@cloudera.com>